Move test on openapi example to lib.rs
